Added new classes v1.2

+++ src/com/sobczak/artur/EmployedPeople.java
@@ -3,23 +3,15 @@ package com.sobczak.artur;
 public class EmployedPeople {
 
     private String name, secondName, surname, address;
-    private int age, pesel, phone;
+    private long age, pesel, phone;
 
-//    public EmployedPeople(String name, String secondName, String surname, String address, int age, int pesel, int phone){
-//        this.name = name;
-//        this.secondName = secondName;
-//        this.surname = surname;
-//        this.address = address;
-//        this.age = age;
-//        this.pesel = pesel;
-//        this.phone = phone;
-//    }
 
     public String getName() {
         return name;
     }
 
     public void setName(String name) {
+        // TODO add exception
         this.name = name;
         System.out.println("Name has added!");
     }
@@ -29,6 +21,7 @@ public class EmployedPeople {
     }
 
     public void setSecondName(String secondName) {
+        // TODO add exception
         this.secondName = secondName;
         System.out.println("Second name has added!");
     }
@@ -38,6 +31,7 @@ public class EmployedPeople {
     }
 
     public void setSurname(String surname) {
+        // TODO add exception
         this.surname = surname;
         System.out.println("Surname has added!");
     }
@@ -47,33 +41,37 @@ public class EmployedPeople {
     }
 
     public void setAddress(String address) {
+        // TODO add exception
         this.address = address;
         System.out.println("Addres has added!");
     }
 
-    public int getAge() {
+    public long getAge() {
         return age;
     }
 
-    public void setAge(int age) {
+    public void setAge(long age) {
+        // TODO add exception
         this.age = age;
         System.out.println("Age has added!");
     }
 
-    public int getPesel() {
+    public long getPesel() {
         return pesel;
     }
 
-    public void setPesel(int pesel) {
+    public void setPesel(long pesel) {
+        // TODO add exception
         this.pesel = pesel;
         System.out.println("Pesel has added!");
     }
 
-    public int getPhone() {
+    public long getPhone() {
         return phone;
     }
 
-    public void setPhone(int phone) {
+    public void setPhone(long phone) {
+        // TODO add exception
         this.phone = phone;
         System.out.println("Phone has added!");
     }
